requirements.txt
python-dotenv
aiogram
uvicorn
fastapi